+++ flex/engines/http_server/handler/graph_db_http_handler.cc
@@ -423,7 +423,84 @@ class adhoc_query_handler : public StoppableHandler {
       std::unique_ptr<seastar::httpd::reply> rep) override {
     auto dst_executor = executor_idx_;
     executor_idx_ = (executor_idx_ + 1) % shard_concurrency_;
-
+    auto& method = req->_method;
+    if (method == "POST") {
+      auto graph_id = trim_slash(req->param.at("graph_id"));
+      if (path.find("vertex") != seastar::sstring::npos) {
+        return executor_refs_[dst_executor]
+            .create_vertex(graph_management_param{
+                std::make_pair(std::move(graph_id), std::move(req->content))})
+            .then_wrapped([rep = std::move(rep)](
+                              seastar::future<admin_query_result>&& fut) mutable {
+              return return_reply_with_result(std::move(rep), std::move(fut));
+            });
+      } else if (path.find("edge") != seastar::sstring::npos) {
+        return executor_refs_[dst_executor]
+            .create_edge(graph_management_param{
+                std::make_pair(std::move(graph_id), std::move(req->content))})
+            .then_wrapped([rep = std::move(rep)](
+                              seastar::future<admin_query_result>&& fut) mutable {
+              return return_reply_with_result(std::move(rep), std::move(fut));
+            });
+      }
+    } else if (method == "GET") {
+      auto graph_id = trim_slash(req->param.at("graph_id"));
+      if (path.find("vertex") != seastar::sstring::npos) {
+        return executor_refs_[dst_executor]
+            .get_vertex(graph_management_query_param{std::make_pair(
+                std::move(graph_id), std::move(req->query_parameters))})
+            .then_wrapped([rep = std::move(rep)](
+                              seastar::future<admin_query_result>&& fut) mutable {
+              return return_reply_with_result(std::move(rep), std::move(fut));
+            });
+      } else if (path.find("edge") != seastar::sstring::npos) {
+        return executor_refs_[dst_executor]
+            .get_edge(graph_management_query_param{std::make_pair(
+                std::move(graph_id), std::move(req->query_parameters))})
+            .then_wrapped([rep = std::move(rep)](
+                              seastar::future<admin_query_result>&& fut) mutable {
+              return return_reply_with_result(std::move(rep), std::move(fut));
+            });
+      }
+    } else if (method == "DELETE") {
+      auto graph_id = trim_slash(req->param.at("graph_id"));
+      if (path.find("vertex") != seastar::sstring::npos) {
+        return executor_refs_[dst_executor]
+            .delete_vertex(graph_management_param{
+                std::make_pair(std::move(graph_id), std::move(req->content))})
+            .then_wrapped([rep = std::move(rep)](
+                              seastar::future<admin_query_result>&& fut) mutable {
+              return return_reply_with_result(std::move(rep), std::move(fut));
+            });
+      } else if (path.find("edge") != seastar::sstring::npos) {
+        return executor_refs_[dst_executor]
+            .delete_edge(graph_management_param{
+                std::make_pair(std::move(graph_id), std::move(req->content))})
+            .then_wrapped([rep = std::move(rep)](
+                              seastar::future<admin_query_result>&& fut) mutable {
+              return return_reply_with_result(std::move(rep), std::move(fut));
+            });
+      }
+    } else if (method == "PUT") {
+      auto graph_id = trim_slash(req->param.at("graph_id"));
+      if (path.find("vertex") != seastar::sstring::npos) {
+        return executor_refs_[dst_executor]
+            .update_vertex(graph_management_param{
+                std::make_pair(std::move(graph_id), std::move(req->content))})
+            .then_wrapped([rep = std::move(rep)](
+                              seastar::future<admin_query_result>&& fut) mutable {
+              return return_reply_with_result(std::move(rep), std::move(fut));
+            });
+      } else if (path.find("edge") != seastar::sstring::npos) {
+        return executor_refs_[dst_executor]
+            .update_edge(graph_management_param{
+                std::make_pair(std::move(graph_id), std::move(req->content))})
+            .then_wrapped([rep = std::move(rep)](
+                              seastar::future<admin_query_result>&& fut) mutable {
+              return return_reply_with_result(std::move(rep), std::move(fut));
+            });
+      }
+    }
     if (path != "/v1/graph/current/adhoc_query" &&
         req->param.exists("graph_id")) {
       // TODO(zhanglei): get from graph_db.
